backend/function_costs/requirements.txt
azure-functions==1.20.0
fastapi==0.115.4
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
azure-identity==1.17.1
# Optional distributed cache:
redis==5.0.7